+++ pandas/core/apply.py
@@ -1251,8 +1251,7 @@ def reconstruct_func(
     Examples
     --------
     >>> reconstruct_func(None, **{"foo": ("col", "min")})
-    (True, defaultdict(<class 'list'>, {'col': ['min']}),
-     ('foo',), array([0], dtype=int64))
+    (True, defaultdict(<class 'list'>, {'col': ['min']}), ('foo',), array([0]))
 
     >>> reconstruct_func("min")
     (False, 'min', None, None)
@@ -1330,8 +1329,7 @@ def normalize_keyword_aggregation(
     Examples
     --------
     >>> normalize_keyword_aggregation({"output": ("input", "sum")})
-    (defaultdict(<class 'list'>, {'input': ['sum']}),
-    ('output',), array([0], dtype=int64))
+    (defaultdict(<class 'list'>, {'input': ['sum']}), ('output',), array([0]))
     """
     from pandas.core.indexes.base import Index
