+++ pandas/core/arrays/sparse/array.py
@@ -359,7 +359,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
     [0, 0, 1, 2]
     Fill: 0
     IntIndex
-    Indices: array([2, 3])
+    Indices: array([2, 3], dtype=int32)
     """
 
     _subtyp = "sparse_array"  # register ABCSparseArray
@@ -642,7 +642,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         --------
         >>> s = SparseArray([0, 0, 1, 0, 2], fill_value=0)
         >>> s.sp_values
-        array([1, 2], dtype=int64)
+        array([1, 2])
         """
         return self._sparse_values
 
@@ -1280,7 +1280,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         [0, 0, 1, 2]
         Fill: 0
         IntIndex
-        Indices: array([2, 3])
+        Indices: array([2, 3], dtype=int32)
 
         >>> arr.astype(SparseDtype(np.dtype('int32')))
         [0, 0, 1, 2]
@@ -1296,7 +1296,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         [nan, nan, 1.0, 2.0]
         Fill: nan
         IntIndex
-        Indices: array([2, 3])
+        Indices: array([2, 3], dtype=int32)
 
         Using a SparseDtype, you can also change the fill value as well.
 
@@ -1305,7 +1305,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         [0.0, 0.0, 1.0, 2.0]
         Fill: 0.0
         IntIndex
-        Indices: array([2, 3])
+        Indices: array([2, 3], dtype=int32)
         """
         if is_dtype_equal(dtype, self._dtype):
             if not copy:
@@ -1358,7 +1358,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         [10, 11, 12]
         Fill: 10
         IntIndex
-        Indices: array([1, 2])
+        Indices: array([1, 2], dtype=int32)
 
         >>> arr.map({0: 10, 1: 11, 2: 12})
         [10, 11, 12]
@@ -1370,7 +1370,7 @@ class SparseArray(OpsMixin, PandasObject, ExtensionArray):
         [10, 11, 12]
         Fill: 10
         IntIndex
-        Indices: array([1, 2])
+        Indices: array([1, 2], dtype=int32)
         """
         # this is used in apply.
         # We get hit since we're an "is_extension_type" but regular extension
