+++ pandas/core/base.py
@@ -1207,13 +1207,13 @@ class IndexOpsMixin(OpsMixin):
         3
 
         >>> ser.searchsorted([0, 4])
-        array([0, 3],  dtype=int64)
+        array([0, 3])
 
         >>> ser.searchsorted([1, 3], side='left')
-        array([0, 2], dtype=int64)
+        array([0, 2])
 
         >>> ser.searchsorted([1, 3], side='right')
-        array([1, 3], dtype=int64)
+        array([1, 3])
 
         >>> ser = pd.Series(pd.to_datetime(['3/11/2000', '3/12/2000', '3/13/2000']))
         >>> ser
